+++ lib/modules/settings/controllers/settings_controller.dart
@@ -1,21 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import '../../../data/services/database_service.dart';
+import '../../../data/services/progress_service.dart';
 import '../../../app/theme/app_theme.dart';
 import 'theme_controller.dart';
 
 class SettingsController extends GetxController {
   final DatabaseService db = Get.find<DatabaseService>();
   final ThemeController themeCtrl = Get.find<ThemeController>();
+  final ProgressService progress = Get.find<ProgressService>();
 
   final selectedLocale = ''.obs;
-  final voucherCount = 0.obs;
 
   @override
   void onInit() {
     super.onInit();
     selectedLocale.value = db.getLocale();
-    voucherCount.value = db.getVouchers().length;
+    progress.reload();
   }
 
   void switchLanguage(String locale) {
@@ -39,7 +40,7 @@ class SettingsController extends GetxController {
 
   Future<void> resetData() async {
     await db.resetAll();
-    voucherCount.value = db.getVouchers().length;
+    progress.reload();
     Get.snackbar('success'.tr, 'settings_reset_success'.tr,
         snackPosition: SnackPosition.BOTTOM);
   }